+++ hmoaserver/src/main/java/hmoa/hmoaserver/config/CacheConfig.java
@@ -1,6 +1,7 @@
 package hmoa.hmoaserver.config;
 
 import com.github.benmanes.caffeine.cache.Caffeine;
+import hmoa.hmoaserver.config.setting.CacheType;
 import org.springframework.cache.CacheManager;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.cache.caffeine.CaffeineCache;
